refactor(core): fold net.rs into capabilities::net
The net adapter trait, ureq backend, NetConfig, and dispatch
function lived at the top level of `aether-substrate-core` while
`capabilities::net::NetCapability` was a thin shell that called into
them — two parallel modules for the same concern. Consolidate the
full stack inside `capabilities::net` so the capability owns its
implementation end-to-end (matching `handle` and `log`, which were
authored that way from the start).

`dispatch_net_mail` is now a private function inside the capability
module instead of `pub(crate)`. `NetConfig::from_env` is the only
external surface; chassis mains import from
`aether_substrate_core::capabilities::net::NetConfig`.

Same approach will follow for io/audio/render in subsequent PRs.
