+++ crates/aether-substrate-desktop/src/main.rs
@@ -1004,7 +1004,7 @@ fn main() -> wasmtime::Result<()> {
     // through the substrate-core APIs explicitly. Substrate-core
     // itself never reads env from now on.
     let hub_url = std::env::var("AETHER_HUB_URL").ok();
-    let net_config = aether_substrate_core::net::NetConfig::from_env();
+    let net_config = aether_substrate_core::capabilities::net::NetConfig::from_env();
     let namespace_roots = aether_substrate_core::io::NamespaceRoots::from_env();
 
     // Shared runtime bring-up: log capture, registry + kind descriptors,
